+++ sdk/go/cloudamqp/integrationMetric.go
@@ -11,6 +11,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v2/go/pulumi"
 )
 
+<<<<<<< HEAD
 // This resource allows you to create and manage, forwarding metrics to third party integrations for a CloudAMQP instance. Once configured, the metrics produced will be forward to corresponding integration.
 //
 // Only available for dedicated subscription plans.
@@ -78,6 +79,8 @@ import (
 // ```sh
 //  $ pulumi import cloudamqp:index/integrationMetric:IntegrationMetric <resource_name> <name>,<instance_id>`
 // ```
+=======
+>>>>>>> 5ed2468... Upgrae to v1.8.3 of the CloudAMQP Terraform Provider
 type IntegrationMetric struct {
 	pulumi.CustomResourceState