+++ sdk/nodejs/integrationMetric.ts
@@ -4,6 +4,7 @@
 import * as pulumi from "@pulumi/pulumi";
 import * as utilities from "./utilities";
 
+<<<<<<< HEAD
 /**
  * This resource allows you to create and manage, forwarding metrics to third party integrations for a CloudAMQP instance. Once configured, the metrics produced will be forward to corresponding integration.
  *
@@ -73,6 +74,8 @@ import * as utilities from "./utilities";
  *  $ pulumi import cloudamqp:index/integrationMetric:IntegrationMetric <resource_name> <name>,<instance_id>`
  * ```
  */
+=======
+>>>>>>> 5ed2468... Upgrae to v1.8.3 of the CloudAMQP Terraform Provider
 export class IntegrationMetric extends pulumi.CustomResource {
     /**
      * Get an existing IntegrationMetric resource's state with the given name, ID, and optional extra